fixed formatting using prettier

+++ contracts/Withdrawable.sol
@@ -5,10 +5,9 @@ import "@openzeppelin/contracts/access/Ownable.sol";
 import "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 
 abstract contract Withdrawble is Ownable {
-
     // for safety reason, if founds somehow stuck
     function withdraw(IERC20 token) external onlyOwner {
-        uint balance = token.balanceOf(address(this));
+        uint256 balance = token.balanceOf(address(this));
         token.transfer(msg.sender, balance);
     }
 }
